test new obfuscation :)

+++ controllers/uploadController.js
@@ -204,12 +204,11 @@ uploadsController.delete = async (req, res) => {
   const file = await db.table('files')
     .where('id', id)
     .where(function () {
-      if (!utils.isAdmin(user.username)) {
-        this.where('userid', user.id)
-      }
+      if (!utils.isAdmin(user.username)) this.where('userid', user.id)
+      
     })
     .first()
-
+  if(!file) return res.json({ success: false, description: 'No file found' })
   try {
     await uploadsController.deleteFile(file.name)
     await db.table('files').where('id', id).del()
